unified-dashboard/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
slack-sdk==3.26.1
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
jira==3.5.2
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
websockets==12.0
apscheduler==3.10.4
email-validator==2.1.0
httpx==0.25.2
